+++ packages/giselle-engine/src/core/runs/utils.ts
@@ -1,3 +1,4 @@
+import { dataMod } from "@giselle-sdk/data-mod";
 import { Run, type RunId } from "@giselle-sdk/data-type";
 import type { Storage } from "unstorage";
 
@@ -21,6 +22,22 @@ export async function setRun({
 	}
 }
 
+function parseAndMod(runLike: unknown, mod = false) {
+	const parseResult = Run.safeParse(runLike);
+	if (parseResult.success) {
+		return parseResult.data;
+	}
+	if (mod) {
+		throw parseResult.error;
+	}
+
+	let modData = runLike;
+	for (const issue of parseResult.error.issues) {
+		modData = dataMod(modData, issue);
+	}
+	return parseAndMod(modData, true);
+}
+
 export async function getRun({
 	storage,
 	runId,
@@ -32,5 +49,5 @@ export async function getRun({
 	if (run == null) {
 		return undefined;
 	}
-	return Run.parse(run);
+	return parseAndMod(run);
 }
